Sema: Move preCheckExpression() into its own file
